add: useWindowResize hook
